+++ packages/ricos-editor/web/src/utils/errorBlocksRemover.ts
@@ -45,7 +45,7 @@ const removeEntitiesWithErrors = (entityMap: RicosEntityMap): RicosEntityMap => 
     const newData = complexPluginHandlers[entityType]?.(data, newEntityMap[entityKey].data);
     if (newData) {
       newEntityMap[entityKey].data = newData;
-    } else if (data.error) {
+    } else if (data?.error) {
       // eslint-disable-next-line fp/no-delete
       delete newEntityMap[entityKey];
     }
@@ -57,7 +57,7 @@ const isErroredBlock = (block: RicosContentBlock, entityMap: RicosEntityMap): bo
   const { entityRanges = [], type } = block;
   const entityKey = entityRanges[0]?.key;
   const { data } = entityMap[entityKey] || {};
-  return type === 'atomic' && data.error;
+  return type === 'atomic' && !!data?.error;
 };
 
 export default function errorBlocksRemover(contentState: DraftContent): DraftContent {