+++ e2e/tests/media.e2e.js
@@ -5,8 +5,9 @@ import {
   PLUGIN_TOOLBAR_BUTTONS,
   GALLERY_SETTINGS,
   GALLERY_IMAGE_SETTINGS,
-  IMAGE_SETTINGS,
   GIPHY_PLUGIN,
+  VIDEO_SETTINGS,
+  SETTINGS_PANEL,
 } from '../cypress/dataHooks';
 import { DEFAULT_DESKTOP_BROWSERS } from './settings';
 import { usePlugins, plugins } from '../cypress/testAppConfig';
@@ -71,14 +72,14 @@ describe('plugins', () => {
 
     it('render image toolbar and settings', function() {
       cy.loadRicosEditorAndViewer('images');
-      cy.openImageSettings();
-      cy.get(`[data-hook=${IMAGE_SETTINGS.PREVIEW}]:first`);
+      cy.openPluginToolbar(PLUGIN_COMPONENT.IMAGE);
+      cy.openSettings();
       cy.eyesCheckWindow({ tag: this.test.title + ' - settings', target: 'window', fully: false });
       cy.addImageTitle();
       cy.eyesCheckWindow(this.test.title + ' - add image title');
       cy.editImageTitle();
       cy.eyesCheckWindow(this.test.title + ' - in plugin editing');
-      cy.openImageSettings(false).deleteImageTitle();
+      cy.openSettings().deleteImageTitle();
       cy.eyesCheckWindow(this.test.title + ' - delete image title');
       cy.addImageLink();
       cy.eyesCheckWindow(this.test.title + ' - add a link');
@@ -301,6 +302,23 @@ describe('plugins', () => {
         cy.waitForGalleryImagesToLoad();
         cy.eyesCheckWindow(this.test.parent.title + ' - delete all items');
       });
+
+      context('settings', () => {
+        it('should disable gallery expand', () => {
+          cy.loadRicosEditorAndViewer('gallery');
+          cy.openPluginToolbar(PLUGIN_COMPONENT.GALLERY);
+          cy.openSettings(['ADV_SETTINGS']);
+          cy.eyesCheckWindow();
+          cy.get(`[data-hook=${GALLERY_SETTINGS.GALLERY_EXPAND_TOGGLE}]`).click();
+          cy.get(`[data-hook=${SETTINGS_PANEL.DONE}]`).click();
+          cy.wait(200);
+          cy.get(`[data-hook=${PLUGIN_COMPONENT.GALLERY}]`)
+            .eq(1)
+            .parent()
+            .click();
+          cy.eyesCheckWindow();
+        });
+      });
       // TODO: title and link image tests
       // // eslint-disable-next-line mocha/no-skipped-tests
       // it.skip('allow to add a title', function() {
@@ -347,6 +365,16 @@ describe('plugins', () => {
         .type('Will this fix the flakiness?');
       cy.waitForVideoToLoad();
       cy.eyesCheckWindow(this.test.title);
+    });
+
+    it('should toggle download option', () => {
+      cy.loadRicosEditorAndViewer('video');
+      cy.openPluginToolbar(PLUGIN_COMPONENT.VIDEO);
+      cy.openSettings();
+      cy.eyesCheckWindow();
+      cy.get(`[data-hook=${VIDEO_SETTINGS.DOWNLOAD_TOGGLE}]`).click();
+      cy.eyesCheckWindow();
+      cy.get(`[data-hook=${SETTINGS_PANEL.DONE}]`).click();
     });
   });