+++ CHANGELOG.md
@@ -20,6 +20,8 @@
 ### :bug: Bug Fix
 - `gallery`
   - [#1705](https://github.com/wix-incubator/rich-content/pull/1705) fix gallery layout styles default and thumbnails ratio
+- `ricos-editor`
+  - [#2336](https://github.com/wix-incubator/rich-content/pull/2336) blocks with errors filter fix
 
 </details>
 <hr/>