+++ CHANGELOG.md
@@ -21,7 +21,6 @@
 
 - `plugin-image`
   - [#2176](https://github.com/wix-incubator/rich-content/pull/2176) update image settings modal to have download and expand options using toggle buttons
-- `plugin-gallery`
 
 ### :bug: Bug Fix
 - `gallery`
@@ -38,6 +37,7 @@
 
 - `plugin-video`
   - [#2331](https://github.com/wix-incubator/rich-content/pull/2331) added new settings panel for custom videos with download toggle
+- `plugin-gallery`
   - [#2096](https://github.com/wix-incubator/rich-content/pull/2096) update gallery settings modal to have download and expand options using toggle buttons
 
 ### :house: Internal