+++ CHANGELOG.md
@@ -16,15 +16,51 @@
     Changes that have landed in master but are not yet released.
     Click to see more.
   </summary>
-	
+	 
 ### :bug: Bug Fix
-
+- `gallery`
+  - [#1705](https://github.com/wix-incubator/rich-content/pull/1705) fix gallery layout styles default and thumbnails ratio
+- `ricos-editor`
+  - [#2336](https://github.com/wix-incubator/rich-content/pull/2336) blocks with errors filter fix
 - `file-upload`
-  - [#2330](https://github.com/wix/ricos/commit/2330) add responsive style when used inside innerrce  
-
+  - [#2330](https://github.com/wix/ricos/commit/2330) add responsive style when used inside innerrce
 
 </details>
 <hr/>
+
+## 8.28.0 (Mar 31, 2021)
+
+### :rocket: New Feature
+
+- `plugin-video`
+  - [#2331](https://github.com/wix-incubator/rich-content/pull/2331) added new settings panel for custom videos with download toggle
+- `plugin-gallery`
+  - [#2096](https://github.com/wix-incubator/rich-content/pull/2096) update gallery settings modal to have download and expand options using toggle buttons
+
+### :house: Internal
+
+- `unsupported-blocks/accordion/table/spoiler`
+  - [#2307](https://github.com/wix/ricos/pull/2307) clean disable-right-click prop from unnecessary components
+- `plugin-table`
+  - [#2303](https://github.com/wix/ricos/pull/2303) add table bi
+- `ricos-content`
+  - [#2237](https://github.com/wix-incubator/rich-content/pull/2237) ContentAPI infra added
+- `general`
+  - [#2237](https://github.com/wix-incubator/rich-content/pull/2237) tsconfig: strictPropertyInitialization; eslint: lines-between-class-members rules
+
+### :bug: Bug Fix
+
+- `image`
+  - [#2327](https://github.com/wix-incubator/rich-content/pull/2327) aria-hidden attribute fixed
+- `plugin-spoiler`
+  - [#2308](https://github.com/wix/ricos/pull/2308) increase blur for media, change spoiler button position in formatting toolbar, change spoiler icon
+- `plugin-link-preview`
+  - [#2324](https://github.com/wix/ricos/pull/2324) remove large space from plugin bottom
+
+### :nail_care: Polish
+
+- `common`
+  - [#2327](https://github.com/wix-incubator/rich-content/pull/2327) TextButtonMapping and InlineToolbarButton types improved
 
 ## 8.27.8 (Mar 30, 2021)
 
@@ -38,8 +74,6 @@
 - `plugin-spoiler`
   - [#2318](https://github.com/wix/ricos/pull/2318) filter spoilered media from preview
 
-## 8.27.6 (Mar 25, 2021)
-
 ### :nail_care: Polish
 
 - `plugin-headings`
@@ -51,7 +85,6 @@
 
 - `color-picker`
   - [#2159](https://github.com/wix-incubator/rich-content/pull/2159) update color-picker style for mobile devices
-
 ### :house: Internal
 
 - `editor-common`